split actor that switches channel and pins mirror
